refactor: rename files (tsx > ts), add index files
