+++ src/interfaces/Links.interface.ts
@@ -1,0 +1,4 @@
+export default interface LinkData {
+  path: string;
+  name: string;
+}
