--- src/interfaces/Links.interface.tsx
@@ -1,4 +1,0 @@
-export interface LinkData {
-  path: string;
-  name: string;
-}
