+++ src/interfaces/User.interface.ts
@@ -1,4 +1,4 @@
-export interface UserInfo {
+export default interface UserInfo {
   id: number;
   avatar: string;
   first_name: string;
