+++ src/interfaces/index.ts
@@ -1,0 +1,5 @@
+export { default as CountriesInfo } from './Countries.interface';
+export { default as FormRefs } from './FormRefs.interface';
+export { default as LinkData } from './Links.interface';
+export { default as PlacesInfo } from './Places.interface';
+export { default as UserInfo } from './User.interface';
