+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export { default as HomePage } from './HomePage';
+export { default as AboutPage } from './AboutPage';
+export { default as PostPage } from './PostPage';
+export { default as ErrorPage } from './ErrorPage/ErrorPage';
